requirements.txt
# Core dependencies for code-pecker project
openai>=1.0.0
httpx>=0.24.0
requests>=2.28.0
pathspec>=0.10.0
neo4j>=5.0.0
python-dotenv>=1.0.0
